employee checks

+++ db/migrate/20131119092954_create_admin_checks.rb
@@ -2,8 +2,8 @@ class CreateAdminChecks < ActiveRecord::Migration
   def change
     create_table :admin_checks do |t|
       t.date :date
-      t.decimal :year
-      t.decimal :month
+      t.integer :year
+      t.integer :month
 
       t.timestamps
     end
